# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kabadiwala Connect API",
    "description": "Smart India Hackathon 2026 (PS 26229) - Informal to Formal E-Waste Bridge API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/prices/{category}": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Get Price By Category",
        "operationId": "get_price_by_category_api_v1_prices__category__get",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prices/{category}/history": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Get Price History",
        "operationId": "get_price_history_api_v1_prices__category__history_get",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots": {
      "post": {
        "tags": [
          "Lots"
        ],
        "summary": "Create Scrap Lot",
        "operationId": "create_scrap_lot_api_v1_lots_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapLotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Lots"
        ],
        "summary": "List Collector Lots",
        "operationId": "list_collector_lots_api_v1_lots_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScrapLotResponse"
                  },
                  "title": "Response List Collector Lots Api V1 Lots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}": {
      "get": {
        "tags": [
          "Lots"
        ],
        "summary": "Get Lot By Id",
        "operationId": "get_lot_by_id_api_v1_lots__lot_id__get",
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapLotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLotRequest": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "condition": {
            "type": "string",
            "title": "Condition",
            "default": "GOOD"
          },
          "weight_kg": {
            "type": "number",
            "maximum": 5000.0,
            "exclusiveMinimum": 0.0,
            "title": "Weight Kg"
          },
          "unit_price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Unit Price"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "default": 0.95
          },
          "photo_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Urls",
            "default": []
          },
          "photo_hashes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Hashes",
            "default": []
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "client_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Uuid"
          },
          "matched_recycler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Recycler Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "category",
          "weight_kg",
          "unit_price"
        ],
        "title": "CreateLotRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PriceRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "category_name_hi": {
            "type": "string",
            "title": "Category Name Hi"
          },
          "category_name_en": {
            "type": "string",
            "title": "Category Name En"
          },
          "base_rate_per_kg": {
            "type": "number",
            "title": "Base Rate Per Kg"
          },
          "min_rate_per_kg": {
            "type": "number",
            "title": "Min Rate Per Kg"
          },
          "max_rate_per_kg": {
            "type": "number",
            "title": "Max Rate Per Kg"
          },
          "cpcb_floor_price": {
            "type": "number",
            "title": "Cpcb Floor Price"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "trend_percentage": {
            "type": "number",
            "title": "Trend Percentage"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "category_name_hi",
          "category_name_en",
          "base_rate_per_kg",
          "min_rate_per_kg",
          "max_rate_per_kg",
          "cpcb_floor_price",
          "trend",
          "trend_percentage",
          "unit",
          "last_updated"
        ],
        "title": "PriceRecordResponse"
      },
...
      "ScrapLotResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "collector_id": {
            "type": "string",
            "title": "Collector Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "condition": {
            "type": "string",
            "title": "Condition"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "estimated_payout": {
            "type": "number",
            "title": "Estimated Payout"
          },
          "final_payout": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Payout"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "client_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Uuid"
          },
          "matched_recycler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Recycler Id"
          },
          "qr_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Token"
          },
          "fallback_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Code"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "photo_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Urls",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "collector_id",
          "category",
          "condition",
          "weight_kg",
          "unit_price",
          "estimated_payout",
          "confidence_score",
          "status",
          "verification_status",
          "created_at",
          "updated_at"
        ],
        "title": "ScrapLotResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}